requirements.txt
ckantoolkit>=0.0.3
# (canada fork only): we use a custom frictionless-py fork
#frictionless==5.0.0b9
markupsafe==2.0.1
tableschema
# (canada fork only): servers cannot call to github
#-e git+https://github.com/ckan/ckanext-scheming.git#egg=ckanext-scheming
